refactor courier model
