+++ src/main/java/ru/yandex/practicum/scooter/api/model/courier/LoginCourierResponse.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.scooter.api.model;
+package ru.yandex.practicum.scooter.api.model.courier;
 
 public class LoginCourierResponse {
     private int id;
